Do a bit of custom lexing in the test class
Instead of manually adding each space token, have the TokenChecker
do that additional work. It's not too hard, and it results in nicer-
looking tests. But because it's a bit of added complexity, I'm adding
a `.raw(...)` checker function, which acts as a sanity check on the
test itself (no DenterHelper work at that point).