use colon instead of braces for method invocation
Also use it for method declaration. We can do this because "upcsating"
(called "component casting" for now) uses @ instead of the colon as I
originally imagined. This means that invocations are unambiguous.
Something like:

    foo bar: Baz

is unambiguously a method invocation of "foo" with args "bar" and "Baz",
as opposed to one arg "bar" of type "Baz." The latter would be:

    foo bar @ Baz

One almost-wrinkle is in the method declaration, since the args do use
colons to declare types. But that's unambiguous because each arg must
start with a var name (ie, starts lowercase). So something like:

    foo arg0 : Bar

is unambiguously a method with one arg, arg0 of type Bar, while

    foo arg0: arg1

is unambiguously a method with two args, both of which have no type.
There is a bit of ugliness, in that the full form is

    foo arg0: Bar: arg1: Baz

But now that I think of it, I don't even need the colon between arg0 and
Bar...