add class to invoke the token literals on contexts
For instance, if there's a method `COLON()` on some context subclass,
invoke it.

This is a silly hack to get the code coverage numbers up -- not for the
sake of "making quota," but so I can focus on real coverage issues.
Ideally I wouldn't even see these methods, as they correspond to string
literals in the grammar and thus don't have any useful info. But see
http://stackoverflow.com/questions/18768905 as to why these methods are
there (tl;dr: antlr4 gives us the full parse tree, and lets us figure
out which parts we want).

This is done via a helper class and a static reference in
EffesParserTest because the instrumentation takes a non-trivial amount
of time; without this short-circuiting, the code coverage takes a while!