split single-line and multiline exprs
Case exprs are currently the only multiline exprs we have. This
distinction is important for a few reasons. Firstly, it lets us make the
NL after `RETURN expr NL` mandatory, which prevents the user from typing
things like `return 1 return 2` (which would previously be interpreted
as two statements, since the NL after the first was optional). Ditto for
assignment statements. Secondly, it prevents the user from typing an
expression which is a multiline directly added to something as in
the file _casePlusInt.ef. This wouldn't cause an ambiguity, but it is
confusing.

Also, improve EffesParserTest so that it complains about unconsumed
input. This is important because most of the rules we're testing don't
include the EOF.