Add test of "one two three four"
Ideally this wouldn't parse, since it's a bit subtle. But it is parsed,
and it's interpreted as `one two (three four)`, where:

- `one` is some expression
- `two` is the name of a method on `one`
- `(three four)` is the only arg to that method, and consists of:
  - an expression `three`
  - the name of a method `four` on the object `three`

I almost wonder if I should create a rule to catch this case
specifically, so that the interpreter can complain about it. For now, I
won't; I'll let the type checker handle it (since it probably wouldn't
check out).