re-introduce methodName for to allow overloads
For instance, `a + b` can now be an AddOrSubExpr or a MethodInvoke,
with the former taking precedence. This allows the user to use
multi-arg syntax for overloads, though at the cost of operator
precedence rules. This is especially useful for things like
division, where you want to be able to specify a rounding mode:
`1.0 / 3.0: HALF_DOWN`. It could also be used for comparing
doubles with an epsilon: `1.0 == (0.1 * 10): Absolute(0.1)`.