handle "self-referencing" types
For instance, if `Foo` has a method `bar` with a return type of
`Foo`. What I'll do is to first generate the `EfMethodMeta`s
with a placeholder of `owningType`. When these metas are finally
passed to the `EfType` that will own them, that type will map
them to replace `owningType` with `this`. It's not perfect, but
it should work.