small cleanups
I'm still not happy with this, but I think it's okay for now. The
issue is that to write the rv "cleanly," we need to know how many
args we have. That's known by the caller of openFrame(), but not
by the return statement; if this were real assembly, the return
statement would know it (but it wouldn't care, actually, cause it
would use a register).

The only major downside to this whole approach is that I can't
efficiently pop the stack by doing a memcpy of nulls to clear the
local vars, because I need to check them as I go to make sure I
haven't hit the esp. But I can't do that anyway since I'm using
ArrayList instead of a native array (for now), so it's no biggie.
I'm going to move on to bigger fish, and revisit this later if
I want.

Oh, another downside to the esp approach rather than the nArgs
approach is that I can't cache values as easily.