simpler, cleaner CallStack
The frame's "anchor" is now a FrameInfo, which contains a few
pieces of info: the number of args, the index of the previous
frame's FrameInfo, and the last index of the previous frame (which
includes the rv slot for this frame). This is more info than a
standard call frame would have (it would only have the previous
sp), but that's because a standard call frame is maintained by
each calling method, not by some CallStack-type object. This
approach is less efficient, but efficiency is already out the
window! This lets each node in the execution tree (e.g. the
return statement) work independently of the context it's in. For
instance, a return statement doesn't need to know how many args
its containing method has.