tests, and a fix of disjunctive types
Case patterns have a result type which is a disjunction of all of
their branches. If all branches return the same type, you get a
disjunction of one type. This should instead be a SimpleType.
So, hide the DisjunctiveType ctor behind a factory method that
checks if the disjunction is only of one type; if it is, return
that type plain.