remove CaseMatcher and options
I had all this framework up for different kinds of alternatives,
when there's only one real kind plus one error. This was in
anticipation of conjunctive types, but I'll cross that bridge
when I get there.

The error handling is now done by returning null, and excluding
those from the AST. This may seem like avoidance, the code does
report an error in that case -- so really it means that the error
is logged and we don't log a subsequent error when trying to
validate that alternative, which we know is invalid.