add support for invoking no-arg methods
This requires a bit of ugliness, since it means that an expression
"foo" could be either a method or a variable. So first of all, I
have to unify the grammar rules. I then dispatch from there based
on whether (a) the thing could be a var and (b) there is a var
by that name. One consequence is that if there isn't a var or
method by that name, we can't just say we didn't find a var; we
have to say we didn't find a var or a method, since the grammar
says the unrecognized name could be either.