refactor Block a bit
The required type goes to the ctor now instead of the validate
method. This lets us build statements "outside-in" instead of
"inside-out." That is, rather than building a block and then asking
whether it validates for a given method, we build a block for a
given method and then ask if it validates.

This is useful because it mirrors the contract for Node
(specifically, that validate takes only the CompileErrs arg) and
thus lets us create a BlockStatement, or even make Block a
subclass of Node.