tweaks to IR printout
We now have the concept of a node which is elided from the
state output. These nodes are wrappers that we don't want to print
in the tree; we just want their wrapped values. The two examples
are Blocks (which wrap a list of Statements) and statement method
invocations (which wrap expression method invocations).