start stubbing out the generics declrs
I've added it to the grammar, but haven't handled it anywhere. But to
make sure I'd handle it correctly in all places, I did a couple things.
Firstly, I renamed `singleType` to `singleTypeA`, so that the (Java)
compiler would catch all of the usages. In each one of those, I
"handled" the generics via a stub on `EfType` (for lack of a better
place to put it), which simply checks that there is no generic declared.
This ensures (a) that I caught all of the usages, and (b) that the diff
reflects that (each change of `singleType` to `singleTypeA` should be
accompanied by a call to the stubbed handler). As a next step, I'll
rename `singleTypeA` back to `singleType`, and also make the handler
stricter by making it always fail. This will cause all tests to fail
until their respective code gets rid of the stubbed handler call and
replaces it with something real.