check that builtin types not declared in userland
To do this, we wrap each compilation unit in a Source, which can have
additional metadata. You can imagine it eventually having other
metadata, like the file name

Tests now fail, because _prefix.ef is a userland file but tries to
declare Int. I'm doing this intentionally, so show that the code works.
:)