fix infinite recursion
ok, gotta admit -- I'm not _entirely_ sure why this happens, or
rather why it only happens when the type recursion goes through an
alias. `type One(prev: BoolInt)` triggers a stack overflow, while
`type One(prev: One | Zero` does not. But, that's the case, and I've
fixed it, so yay.