fix NPE in caseExpression
If there's a parsing error, the ExprBlockContext may be null, which
results in an NPE if we always just to `alt.exprBlock().expr()`. So
we only return the `exprBlock()` part, and then let the caller handle
the case where it's null, and otherwise have the "compile" lambda do
the unrolling to get at the `expr()`.