record whether a large-domain value is exhaustive
A value is exhaustive if the match is against the type, not just a
literal of that type. For instance, "IntValue" is exhaustive, but "10"
is not.

We could potentially do more work to track the values -- for
instance, to track that if you have an IntValue and you do a `case of
10` against it, that the result is "IntValue that's not 10." That might
be nice for future work, but for now we don't do it. There's even a test
of that in match-large-exhaustively.ef.